admission: segment test files by central type
Pure code movement change. Carve out
{io_load_listener,tokens_linear_model}_test.go using the same non-test
partitioning scheme as earlier commits. Within granter_test.go, move the
supporting types further below in the file.

Release note: None
