+++ pkg/util/admission/BUILD.bazel
@@ -40,7 +40,9 @@ go_test(
     srcs = [
         "disk_bandwidth_test.go",
         "granter_test.go",
+        "io_load_listener_test.go",
         "store_token_estimation_test.go",
+        "tokens_linear_model_test.go",
         "work_queue_test.go",
     ],
     data = glob(["testdata/**"]),
